CON-4802 Always store remote categories
